Follow-up to [38dc27bd1d0ecd682aafde]: Tk does not support <Button-6> nor <Button-7> events. This commit adds hardware-support for buttons 6-9 on MacOSX, and handles masking correct for all buttons 6-9. Not complete yet, but this is as far as we can go without needing a TIP.

+++ generic/tkGrab.c
@@ -872,7 +872,7 @@ TkPointerEvent(
 	} else {
 	    if (eventPtr->xbutton.button != AnyButton &&
 		    ((eventPtr->xbutton.state & ALL_BUTTONS)
-		    == (unsigned int)TkGetButtonMask(eventPtr->xbutton.button))) {
+		    == TkGetButtonMask(eventPtr->xbutton.button))) {
 		ReleaseButtonGrab(dispPtr);			/* Note 4. */
 	    }
 	}
